Set cookie using JS
We can’t set the cookie if a non-inline form was displayed, as we don’t know if it will be displayed until the footer of the site is loaded. setcookie() is too late, so JS has to be used to check for `script[data-kit-limit-per-session]` and set the cookie that way.